Rename DirectBitmapCapturer to BitmapCapturer. Remove DirectBitmap.
