--- SharedLib/DirectBitmap/IDirectBitmapProvider.cs
@@ -1,7 +1,0 @@
-﻿namespace SharedLib
-{
-    public interface IDirectBitmapProvider
-    {
-        DirectBitmap DirectBitmap { get; }
-    }
-}
